chess_ai_commentary/requirements.txt
chess
python-chess
stockfish
google-generativeai
pydub
pygame
torch
torchaudio
numpy
pandas
python-dotenv
gtts
TTS
speechrecognition
pyaudio
moviepy
opencv-python
supabase
fastapi
uvicorn[standard]
pydantic
chess.com
requests